EP-826. Moved methods that prepare data inside BaseTest.
