+++ tvb_epilepsy/tests/test_head_service.py
@@ -1,9 +1,3 @@
-import os
-import numpy
-from tvb_epilepsy.io.h5_reader import H5Reader
-from tvb_epilepsy.base.model.vep.head import Head
-from tvb_epilepsy.base.model.vep.sensors import Sensors
-from tvb_epilepsy.base.model.vep.surface import Surface
 from tvb_epilepsy.service.head_service import HeadService
 from tvb_epilepsy.base.constants.config import Config
 from tvb_epilepsy.tests.base import BaseTest
@@ -12,15 +6,6 @@ from tvb_epilepsy.tests.base import BaseTest
 class TestHeadService(BaseTest):
     head_service = HeadService()
     config = Config()
-
-    def _prepare_dummy_head(self):
-        reader = H5Reader()
-        connectivity = reader.read_connectivity(os.path.join(self.config.input.HEAD, "Connectivity.h5"))
-        cort_surface = Surface([], [])
-        seeg_sensors = Sensors(numpy.array(["sens1", "sens2"]), numpy.array([[0, 0, 0], [0, 1, 0]]))
-        head = Head(connectivity, cort_surface, sensorsSEEG=seeg_sensors)
-
-        return head
 
     def test_select_sensors_power(self):
         head = self._prepare_dummy_head()
